+++ lib/shared/src/chat/recipes/git-log.ts
@@ -3,6 +3,7 @@ import path from 'path'
 
 import { MAX_RECIPE_INPUT_TOKENS } from '../../prompt/constants'
 import { truncateText } from '../../prompt/truncation'
+import { newInteraction } from '../prompts/utils'
 import { Interaction } from '../transcript/interaction'
 
 import { Recipe, RecipeContext, RecipeID } from './recipe'
@@ -80,16 +81,12 @@ export class GitHistory implements Recipe {
 
         const promptMessage = `Summarize these commits:\n${truncatedGitLogOutput}\n\nProvide your response in the form of a bulleted list. Do not mention the commit hashes.`
         const assistantResponsePrefix = `Here is a summary of recent changes:\n${truncatedLogMessage}`
-        return new Interaction(
-            { speaker: 'human', text: promptMessage, displayText: rawDisplayText, source },
-            {
-                speaker: 'assistant',
-                prefix: assistantResponsePrefix,
-                text: assistantResponsePrefix,
-                source,
-            },
-            Promise.resolve([]),
-            []
-        )
+        return newInteraction({
+            text: promptMessage,
+            displayText: rawDisplayText,
+            source,
+            assistantPrefix: assistantResponsePrefix,
+            assistantText: assistantResponsePrefix,
+        })
     }
 }