+++ lib/ui/src/chat/CodeBlocks.tsx
@@ -15,6 +15,8 @@ import {
 import styles from './CodeBlocks.module.css'
 
 interface CodeBlocksProps {
+    inProgress: boolean
+
     displayText: string
 
     copyButtonClassName?: string
@@ -23,7 +25,12 @@ interface CodeBlocksProps {
     copyButtonOnSubmit?: CodeBlockActionsProps['copyButtonOnSubmit']
     insertButtonOnSubmit?: CodeBlockActionsProps['insertButtonOnSubmit']
 
+    metadata?: CodeBlockMeta
+}
+
+export interface CodeBlockMeta {
     source?: string // the name of the executed command that generated the code
+    requestID?: string // id of the request that generated the code
 }
 
 function createButtons(
@@ -32,7 +39,7 @@ function createButtons(
     copyButtonOnSubmit?: CodeBlockActionsProps['copyButtonOnSubmit'],
     insertButtonClassName?: string,
     insertButtonOnSubmit?: CodeBlockActionsProps['insertButtonOnSubmit'],
-    source?: string
+    metadata?: CodeBlockMeta
 ): HTMLElement {
     const container = document.createElement('div')
     container.className = styles.container
@@ -57,7 +64,7 @@ function createButtons(
         CopyCodeBlockIcon,
         codeBlockActions,
         copyButtonClassName,
-        source
+        metadata
     )
     buttons.append(copyButton)
 
@@ -71,7 +78,7 @@ function createButtons(
                 InsertCodeBlockIcon,
                 codeBlockActions,
                 insertButtonClassName,
-                source
+                metadata
             )
         )
 
@@ -83,7 +90,7 @@ function createButtons(
                 SaveCodeBlockIcon,
                 codeBlockActions,
                 insertButtonClassName,
-                source
+                metadata
             )
         )
     }
@@ -95,7 +102,6 @@ function createButtons(
 
 /**
  * Creates a button to perform an action on a code block.
- *
  * @returns The button element.
  */
 function createCodeBlockActionButton(
@@ -108,7 +114,7 @@ function createCodeBlockActionButton(
         insert?: CodeBlockActionsProps['insertButtonOnSubmit']
     },
     className?: string,
-    source?: string
+    metadata?: CodeBlockMeta
 ): HTMLElement {
     const button = document.createElement('button')
 
@@ -123,7 +129,7 @@ function createCodeBlockActionButton(
             button.innerHTML = CheckCodeBlockIcon
             navigator.clipboard.writeText(text).catch(error => console.error(error))
             button.className = classNames(styleClass, className)
-            codeBlockActions.copy(text, 'Button', source)
+            codeBlockActions.copy(text, 'Button', metadata)
             setTimeout(() => (button.innerHTML = iconSvg), 5000)
         })
     }
@@ -135,10 +141,10 @@ function createCodeBlockActionButton(
 
     switch (type) {
         case 'insert':
-            button.addEventListener('click', () => insertOnSubmit(text, false, source))
+            button.addEventListener('click', () => insertOnSubmit(text, false, metadata))
             break
         case 'new':
-            button.addEventListener('click', () => insertOnSubmit(text, true, source))
+            button.addEventListener('click', () => insertOnSubmit(text, true, metadata))
             break
     }
 
@@ -151,11 +157,17 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
     copyButtonOnSubmit,
     insertButtonClassName,
     insertButtonOnSubmit,
-    source,
+    metadata,
+    inProgress,
 }) {
     const rootRef = useRef<HTMLDivElement>(null)
 
     useEffect(() => {
+        // Attach code block actions only when the message is completed
+        if (inProgress) {
+            return
+        }
+
         const preElements = rootRef.current?.querySelectorAll('pre')
         if (!preElements?.length) {
             return
@@ -170,7 +182,7 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
                     copyButtonOnSubmit,
                     insertButtonClassName,
                     insertButtonOnSubmit,
-                    source
+                    metadata
                 )
 
                 // Insert the buttons after the pre using insertBefore() because there is no insertAfter()
@@ -179,7 +191,7 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
                 // capture copy events (right click or keydown) on code block
                 preElement.addEventListener('copy', () => {
                     if (copyButtonOnSubmit) {
-                        copyButtonOnSubmit(preText, 'Keydown')
+                        copyButtonOnSubmit(preText, 'Keydown', metadata)
                     }
                 })
             }
@@ -191,7 +203,8 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
         rootRef,
         copyButtonOnSubmit,
         insertButtonOnSubmit,
-        source,
+        metadata,
+        inProgress,
     ])
 
     return useMemo(