+++ vscode/src/services/telemetry-v2.ts
@@ -53,7 +53,7 @@ const legacyBackcompatLogEventMode: LogEventMode = 'connected-instance-only'
 const debugLogLabel = 'telemetry-v2'
 
 function updateGlobalInstances(updatedProvider: TelemetryRecorderProvider): void {
-    telemetryRecorderProvider?.complete()
+    telemetryRecorderProvider?.unsubscribe()
     telemetryRecorderProvider = updatedProvider
     telemetryRecorder = updatedProvider.getRecorder([
         // Log all events in debug for reference.