+++ vscode/test/benchmark/datasets/typescript/external-package-deep-equal/solution.ts
@@ -1,5 +1,0 @@
-import isEqual from 'deep-equal'
-
-export function areSame(obj1: any, obj2: any): boolean {
-    return isEqual(obj1, obj2)
-}