+++ vscode/test/benchmark/datasets/typescript/utils-import-all-closed-file/solution.ts
@@ -1,6 +1,0 @@
-import { Club } from './types'
-import * as utils from './utils'
-
-export function richClubsTotalBudget(clubs: Club[]): number {
-    return utils.totalBudget(clubs.filter(c => utils.isRich(c)))
-}